Refactor fighters to use a BaseFighter class
