+++ src/index.ts
@@ -5,23 +5,12 @@ export {
 } from "./CommandResponse";
 export { default as History, HistoryJSON } from "./History";
 export { default as Citizen, CitizenJSON } from "./Citizen";
-export {
-  default as CavalryFighter,
-  CavalryFighterJSON
-} from "./CavalryFighter";
-export {
-  default as InfantryFighter,
-  InfantryFighterJSON
-} from "./InfantryFighter";
-export { default as RangedFighter, RangedFighterJSON } from "./RangedFighter";
 export { default as Food, FoodJSON } from "./Food";
 export {
   default as Game,
   GameJSON,
   GameProps,
   GameGenerateProps,
-  Fighter,
-  FighterType,
   Unit
 } from "./Game";
 export { CommandJSON, CommandChildClass } from "./Command";
@@ -32,6 +21,7 @@ export {
   PositionJSON
 } from "./ObjectWithPosition";
 export * from "./commands";
+export * from "./fighters";
 export { default as Team, TeamJSON } from "./Team";
 export { default as Wall, WallJSON } from "./Wall";
 export { default as isValidPosition } from "./utils/isValidPosition";
